+++ lib/puppetserver/ca/action/sign.rb
